clienthandler hass been moved to a separate class
